Refactor: Extract `getModifiedTokenAddresses` to utils file
